### Description
This adds a lock to make sure limiters write are sequential and that we don't override another limiter when calling `Allow` concurrently.

### Testing & Reproduction steps

#### Without any synchronization code in #15467 
``` bash
❯ go test -bench=BenchmarkTestRateLimiterAllowC -count 5 -benchtime=1s  -timeout 6000s -run=XXX ./agent/consul/multilimiter
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/consul/agent/consul/multilimiter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
AC/no_prefill-16         	  556688	      2060 ns/op
AC/no_prefill-16         	  672924	      1964 ns/op
AC/no_prefill-16         	  627190	      1932 ns/op
AC/no_prefill-16         	  615247	      1952 ns/op
AC/no_prefill-16         	  661696	      1907 ns/op
AC/prefill_with_1K_keys-16         	  727677	      1894 ns/op
AC/prefill_with_1K_keys-16         	  611496	      2310 ns/op
AC/prefill_with_1K_keys-16         	  642594	      2043 ns/op
AC/prefill_with_1K_keys-16         	  590719	      1888 ns/op
AC/prefill_with_1K_keys-16         	  625422	      1937 ns/op
AC/prefill_with_10K_keys-16        	  494367	      2436 ns/op
AC/prefill_with_10K_keys-16        	  522981	      2491 ns/op
AC/prefill_with_10K_keys-16        	  529140	      2291 ns/op
AC/prefill_with_10K_keys-16        	  612369	      2356 ns/op
AC/prefill_with_10K_keys-16        	 1000000	      2607 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      3285 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      3626 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      3891 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      4479 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      4352 ns/op
PASS
ok  	github.com/hashicorp/consul/agent/consul/multilimiter	66.699s
```

#### This branch (lock)
```bash
❯ go test -bench=BenchmarkTestRateLimiterAllowC -count 5 -benchtime=1s  -timeout 6000s -run=XXX ./agent/consul/multilimiter
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/consul/agent/consul/multilimiter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
AC/no_prefill-16         	  138690	     10123 ns/op
AC/no_prefill-16         	  147824	      9014 ns/op
AC/no_prefill-16         	  158497	     11059 ns/op
AC/no_prefill-16         	  141631	     10767 ns/op
AC/no_prefill-16         	  149265	     10668 ns/op
AC/prefill_with_1K_keys-16         	  136306	     10550 ns/op
AC/prefill_with_1K_keys-16         	  154232	     12941 ns/op
AC/prefill_with_1K_keys-16         	  135919	     10701 ns/op
AC/prefill_with_1K_keys-16         	  160878	     10177 ns/op
AC/prefill_with_1K_keys-16         	  152521	     10776 ns/op
AC/prefill_with_10K_keys-16        	  172677	     11131 ns/op
AC/prefill_with_10K_keys-16        	  166374	     10944 ns/op
AC/prefill_with_10K_keys-16        	  164202	     10018 ns/op
AC/prefill_with_10K_keys-16        	  151930	     10401 ns/op
AC/prefill_with_10K_keys-16        	  150478	     10966 ns/op
AC/prefill_with_100K_keys-16       	  105031	     11913 ns/op
AC/prefill_with_100K_keys-16       	  137925	     12416 ns/op
AC/prefill_with_100K_keys-16       	  122389	     14066 ns/op
AC/prefill_with_100K_keys-16       	  164812	     13228 ns/op
AC/prefill_with_100K_keys-16       	  163897	     14676 ns/op
PASS
ok  	github.com/hashicorp/consul/agent/consul/multilimiter	69.298s
```